Apply review: make the codec faster than what it replaced, and move it
The first commit accepted a 3.7x encode slowdown as the price of compression.
It was not the price; it was the dictionary.

## The table, not the algorithm

The LZW dictionary was a Map<int, int> keyed on (prefix << 8) | pixel,
allocated fresh per frame. Two things were wrong with that: a hash map for
what is a dense small-integer lookup, and a key sized to a byte when the
palette has four slots - 4096 x 256 entries where 4096 x 4 will do. A flat
Int32List on a palette-width key is 64KB instead of 4MB, which is the
difference between living in cache and not.

Measured on 120 frames of a synthetic Flipper screen, AOT:

  scale   before this PR   first commit   now      output
      1            13 ms          82 ms    8 ms    545 -> 32 KB
      2            79 ms         331 ms   19 ms   2172 -> 70 KB
      4           368 ms        1257 ms   48 ms   8677 -> 158 KB

So the codec is now 55x smaller *and* 7.7x faster than the non-compressing
version it replaced, rather than smaller but slower. Verified byte-identical
to the first commit across 192 configurations of size, scale, content and
frame count, and re-checked against Pillow.

Also: the scaler expands each source row once and copies it down rather than
recomputing a source index per output pixel, and the bit writer accumulates
into a Uint8List rather than a growable List<int>. Neither is worth doing
alone; both are free alongside the dictionary change.

Reusing one table across frames introduces two ways to be wrong that did not
exist when it was allocated per frame - forgetting to clear it between frames,
and forgetting to clear it on a mid-frame reset. Both are now pinned by tests.

## Moved out of a page directory

The codec is pure Dart with no Flutter imports and two unrelated callers, one
of which reached for it as ../../remote/desktop/gif_encoder.dart from the
paint editor. lib/components/codec/ already exists for exactly this and
already holds bm.dart in the same shape.

## Not done, deliberately

- package:image ships a GifEncoder and is already resolved transitively, but
  it hardcodes an 8-bit minimum code size regardless of palette, so it
  produces output 2x larger than this on the same frames, needs palette-backed
  Image objects, and writes a global colour table where this writes per-frame
  local ones. Hand-rolling is right here, which is the answer to the question
  the issue actually asked.
- Inter-frame compression - dirty rectangles plus a transparent index - is
  worth a further 2.4x to 15x on top of this, and is a real feature rather
  than a fix. Filed separately.

Mutation testing: 11 of 11 caught.
